Fix(microphone): Correctly pass ps_api to finalizer
The  class's  was not correctly
accessing the default  implementation when one was not provided
to .

This was because it was accessing the instance variable
directly, which would be  in this case.

This commit fixes this by ensuring that the  accessor method is
used instead, which correctly returns the default
implementation.